training/requirements.txt
setuptools==67.8.0
wheel>=0.38.4
boto3>=1.26.0
ffcv==1.0.2
numpy>=1.21.0
scikit-learn>=1.0.0
tqdm>=4.65.0
opencv-python>=4.8.0
cupy-cuda12x 